backend/requirements.txt
fastapi
strawberry-graphql[fastapi]
sqlalchemy[asyncio]
aiosqlite
pydantic
pydantic[email]
uvicorn[standard]